Make glyphs thicker
The glyphs were widely considered too thin, which even caused people to
try and make them thicker by having Word automatically create a bold
version of them, which has many issues and is an universally hated
feature of Word in the typography community. In this context, it was
essential that a thicker version of the glyphs was created.