Make glyphs case-independent
There were previously different glyphs for upper and lower case letters
in the Latin alphabet, however, the LIBRAS language does not observe
such distinction, so all the upper case letters now use the same glyphs
as the lower case ones. The lower case glyphs were chosen to take the
place of the upper case ones as they are smaller and will better
accommodate the different parts of the glyphs. In fact, the sizing of
the circles in the lower case glyphs is the one that will be used for
all other characters once they are implemented.