Reorganize glyph files
The glyphs are now separate by category. As of this commit, only the
letters are implemented, so the glyphs folder only has one subfolder
with the letters, but more will be added in the future.

The shapes folder was moved to the root of the repository, as it is not
a category of glyphs, but a set of common shapes that are used in the
creation of the glyphs.