Add several common symbols and punctuation marks
Many common symbols and punctuation marks were added to the font. The
characters that were added cover, with the letters and numbers, most of
ASCII. The only printable ASCII characters missing are the carat, tilde
and grave accent. They will be added in the future along with the
accented letters.

Four non-ASCII characters were also added: the left and right double and
single comma quotation marks (“”‘’). These characters are especially
important as they are the ones Word uses when quotation marks are typed.

The symbols added are:
- Ampersand
- Apostrophe
- Asterisk
- At sign
- Colon
- Comma
- Dollar sign
- Equals
- Exclamation mark
- Greater than
- Hash
- Hyphen minus
- Left curly bracket
- Left double quotes
- Left parenthesis
- Left single quotes
- Left square bracket
- Less than
- Percent
- Period
- Pipe
- Plus
- Question mark
- Quotes
- Reverse solidus
- Right curly bracket
- Right double quotes
- Right parenthesis
- Right single quotes
- Right square bracket
- Semicolon
- Solidus
- Underscore