Make number encoding code point offset 0xE1400
The number 10, rather than 0, was originally at the beginning of the
number block (U+E1400–U+E17FF). This had the implication of making the
number code point offset the unusual value 0xE13F6 (0xE1400 - 10) and
causing numbers smaller than 10 to be encoded with a negative offset, at
the end of the previous block. The original encoding had this offset
because it was believed that no number smaller than 10 would ever be
encoded. This may not be the case since it is possible that the decimal
digits will eventually be changed to the usual Arabic numerals to
improve usability in real-world scenarios, and keeping support for the
SER-Libras numeral representation is desirable.

Although this is an incompatible potentially hard-to-debug breaking
encoding change, the encoding has never been formally released, so its
consistency should not be relied on in the first place. This change is
also exceedingly unlikely to cause any unforeseen issues anywhere as the
commiter is the only person known to be working with this encoding. The
only known case this change will require action is in the SER-LIBRAS
app, which will be modified to adapt to this change immediately. This
change is so meaningless, in fact, that the previous encoding was only
used in one unpushed commit in the SER-LIBRAS app repository, that will
be amended, so the change will never even make its way into the app
source history.

The font was modified to reflect this change, the number 10 was moved
from U+E1400 to U+E140A, 11 was moved from U+E1401 to U+E140B, and the
SER-Libras glyphs that are encoded in the place of Arabic numerals
(U+0030–U+0039) were copied to U+E1400–U+E1409, but kept at their Arabic
numeral Unicode code points for now.